Enable/Fix failing tests in cl_spec.lua
Tests in cl_spec.lua were not being executed because they were nested
inside an 'it' block. Tests need to be in a 'describe' block in order to
execute, otherwise they are silently disabled. Nesting an 'it' block
inside another 'it' block should probably be an error. After enabling the
tests, many of the tests were failing.

Tags were not being filtered correctly. Instead of running tests with
one or more tags listed in the --tags option, tests were being excluded
if any of the listed tags were not present.

Need to incorporate pull request #293 in order to fix the compile fail
test. This also addresses issue #273.

According to the tests, the return code for busted should be the number
of errors/failures. However, the return code needs to saturate at 255
since it can only occupy 8-bits.