Fix traceback issue #286
The base output handler was not saving the trace info when it was
notified that an error/failure occurred. The error/failure handler needs
to save the trace information, otherwise the trace information of the
test call site will be used. Furthermore, we do not need to pass in
trace information for test end messages, since the trace information is
already present in the context element at creation. The output handler
formatter just needs to prioritize the passed in debug information over
the context element's trace information since the passed in debug
information will provide more precise location information.