Distinguish between test errors and test failures
A test error occurs when an error is thrown or an assert fails outside
of the test. While test failures only occur when an assert fails inside
a test.

Call setfenv on 'assert' to replace 'error' with 'fail', such that
'fail' throws a special failure table instead of a string. Hence,
assertion failures will throw a failure table and are distinct from
errors. Additionally, the failure table contains the same error string
that would have been thrown with 'error'.