Support 'pending' inside an 'it' block
When 'pending' is called from an 'it' block, the rest of the test should
be skipped and the test should be flagged as pending. This can be
acheived by throwing a 'pending' table when 'pending' is called from an
'it' block, then publishing a 'pending' message when the 'pending' table
is caught. This causes the remainder of the test to be skipped and the
output handler to be notified that the current test is 'pending'. Also,
make sure the --suppress-pending option is honored.