Error when unsupported functions called from test
Calling randomize/describe/it/setup/before_each/after_each/teardown
functions from inside an 'it' block is not valid, so throw an error when
any of these functions are called inside an 'it'. In addition, pending
is not valid inside a 'finally' block. Ensure that the error string
reflects the error level of the caller.