Refactor descriptor executor registration
Currently a debug trace is saved for registered descriptors that do not
match 'file'. However, since the 'file' executor is only called by the
test file loader directly, via busted.executors, before any files are run,
it is sufficient to check that the current context does not have a
parent instead of checking that the descriptor is not 'file'.