Fix use of handler.getFullName in output handlers
The function handler.getFullName() was not being used correctly in the
plainTerminal and utfTerminal output handlers.  The getFullName()
function should be passed a context element as input, but these output
handlers were passing the results of handler.format(), which is not a
context element, and already has a 'name' field that is the result of
handler.getFullName().

The TAP output handler was calling handler.format() on tables that were
already the result of a previous call to handler.format(). In this case,
just use the pre-formatted table directly.

Remove unused local function getFullName() in TAP output handler.