Create support function globals in parent context
When a global is created in a support function (i.e. setup/before_each),
create the global in the current block's parent context so that other
sibling blocks (i.e. 'it' blocks) can access these globals.

Globals created in 'it' blocks are still confined to the current
context/test, and not visible to it's siblings.

Fixes issues #289 and #302.