Fix issue #283 Run duration is always 0.0 seconds
The reported run duration is always 0.0 seconds. This occurs because the
terminal output handlers rely on the base output handler to record the
start and end times of the suite. However, when the terminal output
handler's suiteEnd() function runs, the base output handler's
baseSuiteEnd() function has yet to record the end time. Thus, the suite
end time is still 'nil' when the terminal output handlers are trying to
get the run duration, resulting in a 0 duration.

The solution is to ensure that the base output handler subscribes at the
highest priority so that the suite end time will be recorded before it
is needed by any other output handlers.