Fix issue #250 Weird context popping
At the end of a test suite the last context is popped off the stack
leaving it empty. However, this doesn't seem to be intended as there are
an equal number of pushes/pops. This error occurs because the context
pushes/pops are not done with a stack, but rather relies on the context
parent/child relationships.

Use a stack to properly keep track of context pushes/pops.