Save/Restore global state between test files
When a file context is pushed, save a copy of _G and package.loaded.
Then when the file context is popped, restore the saved values of _G and
package.loaded. This prevents a 'require' in one file from polluting the
global environment of another file. This also allows modules that are
required in one test file to be reloaded in another test file.

This addresses issue #62.