Support running tests standalone
Allow unit tests to be run standalone by just adding a require to the
top of a test file. This is useful if, for some reason, a command line
tool is not viable.

Just add the following require statement to the beginning of a test file:

	require 'busted.runner'()

When the require is present, the test file can be run standalone or with
the busted command line tool.

This addresses issue #294.