Fix error message when pcall occurs before assert
This fixes the case where the file, line number, and error message
reported are incorrect, showing the error message for a previous call
to `error()` that was caught with a `pcall()`.

  describe('Test assert after pcall', function()
    it('should report line number of assert', function()
      local ok, msg = pcall(function() error('error message') end)
      assert(false, 'should report this line number')
    end)
  end)

In the above example, the error message reported would be from the
`pcall()`, instead of the error message from the `assert()`, where the
actual test failure occurred.

The solution is to not replace `error` in the test environment. This can
lead to more inconsistent error messages that may be difficult to trace.
Go back to rewriting the error message to insert the file and line
information for complex error objects.