Display errors/failures in order of execution
This change updates the order that error, failure, and pending messages
are displayed such that they are reported in the same order as when they
occurred. Messages are still grouped according to their type, where all
pending messages are grouped together, all failure messages are grouped
together, etc. This allows for the same exact output messages to be
displayed in the same order each time a test suite is executed with the
same command-line options.