Support loading output handler via Lua search path
This allows an output handler to be loaded from a module found in the
Lua search path. The output handler can still be loaded from a lua or
moonscript file as well. If the output handler specified on the command
line cannot be found in `busted/outputHandlers/`, then attempt to load
the output handler as if it were in the Lua module search path.