Update TAP output handler
This updates the TAP output handler to output the result of each test as
it completes, instead of all test results at the end of a suite. When
the verbose option is enabled, print the test name at test start. Also
update TAP output to report skipped tests and errors occurring outside
of tests.