Fix for running tests standalone
This updates the standalone run mode to use the last child of the root
context as the 'file' descriptor, instead of assuming the first child in
the root context is the 'file' descriptor. A helper script could add
support functions to the root context (e.g. 'setup'/'teardown'/etc), so
the first child may not be a 'file' descriptor, but the last child will
always be a 'file' descriptor since the test file was just loaded
prior to the standalone test file setup for live coding.