Fix unintentional stack trace on exit
This fixes the stack trace on exiting with an error code on Lua 5.3.
On Lua 5.3, `error()` is not silent, so don't throw an `error()` to exit
with error for Lua > 5.2. For Lua versions 5.1 and 5.2 continue to throw
a silent `error()` to allow the Lua state to be closed naturally
whenever possible.