Add waf integration to this patch
This patch allows busted to be installed on a target in the root
system's LUA_PATH so that they can be used by the system. This patch
assumes that pre-requisites such as lua-term, lua_cliargs, mediator.lua,
say, and luassert are all installed.

It uses the build environment variable TARGET_LUA_PATH as the base of
its installation and installs the `busted` script in BINDIR.

Change-Id: I783b47499f3709ef4466588ae90c923249b863cd
Type: New Feature
Reviewed-on: https://gerrit.consumer.garmin.com/code-review/370905
Approve-Change: Harris, James <james.harris@garmin.com>
Reviewed-by: Harris, James <james.harris@garmin.com>
Tested-by: Harris, James <james.harris@garmin.com>