Refactor copying of context in base output handler
This allows for adding new keys to a test context without the need to
change the output handlers in order to copy a new or changed key. We
can't just copy the entire table, otherwise the json output handler
will break. The `env` key and any values that evaluate to functions
cannot be copied.