Allow junit output handler to save output to file
When trying to hook busted's junit output module up to an internal
automated test system it is very challenging to determine what output
was written to stdout by tests, and what output was written to stdout by
tests and what output is part of the test summary in xml. The simplest
way to handle this is to allow junit to accept an argument naming a file
where the xml output should be written via the -Xoutput mechanism. This
patch adds this change to junit.

Change-Id: Iea996783ed62cf707f28557dc9e1ee4055e44642