Fix error reporting for errors in message handlers
When an error occurs in a suite start/end handler, retrieving the trace
information yields another error because there is no file associated
with the code that threw the original error. This then triggers another
error to be thrown, resulting in an infinite loop of error throwing and
error catching. This change fixes this by checking for a valid file
context before accessing the subfields for the file.

When a 'failure' is thrown (from an `assert`) inside a test start/end
handler, the failure and error counters are not incremented if the test
passes. Failures in a test start/end handler should be logged as errors
outside tests, just as failures in support functions are logged as
errors outside tests.

Also use safe_publish for suite reset to catch any errors that may occur
in a suite reset handler.