Fix the random seed sent to suite start
The random seed being sent to suite start is incorrect when multiple
runs are requested. In such cases, the seed passed to suite start is
actually the seed for the previous run and not the current run. This
change passes to suite start the random seed being used for the current
run.