Output error message when test file/dir not found
Previously, if a test file or directory specified on the command-line
is not found, the test file loader would silently fail without reporting
an error. Hence, leading the user to think all requested tests were run,
when, in fact, either an entire test file or directory was skipped. This
updates the test file loader to report an error in the event a requested
file or directory is not found.