fix(action): Avoid GitHub marketplace namespace conflict
Some puppy that claimed the GitHub username in 2011 and then never did
anything with it in a decade is blocking this being published to the
marketplace as 'busted'. I had checked that the market namespace was
open before picking the name but did not remember than *any* username
conflict would be a blocker. This is not ideal but not the end of the
world.